requirements.txt
gradio
pandas
pymupdf
tabulate
google-generativeai
pdfplumber
openpyxl
xlrd
matplotlib
seaborn
langgraph 
langchain 
openai
requests 
python-magic
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
python-docx
streamlit
pytesseract==0.3.13
pdf2image
PyPDF2
sqlalchemy 
psycopg2
